web/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
GitPython>=3.1.40
Pillow>=10.0.0
rembg>=2.0.50
